Extract template data analyzation to the separate class
Do not mix template analyzation logic with initialization logic to avoid using implicit contexts available in the initializer.

Fixes #3587
