+++ flow-server/src/main/java/com/vaadin/flow/component/polymertemplate/TemplateInitializer.java
@@ -16,31 +16,18 @@
 package com.vaadin.flow.component.polymertemplate;
 
 import java.lang.reflect.Field;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
 import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.function.Consumer;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.stream.Stream;
-
-import org.jsoup.nodes.Attribute;
-import org.jsoup.nodes.Node;
-import org.jsoup.select.Elements;
-import org.jsoup.select.NodeVisitor;
 
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.Tag;
+import com.vaadin.flow.component.polymertemplate.TemplateDataAnalyzer.ParserData;
 import com.vaadin.flow.dom.Element;
 import com.vaadin.flow.dom.ShadowRoot;
-import com.vaadin.flow.internal.AnnotationReader;
 import com.vaadin.flow.internal.ReflectTools;
 import com.vaadin.flow.internal.ReflectionCache;
 import com.vaadin.flow.internal.nodefeature.NodeProperties;
@@ -48,7 +35,6 @@ import com.vaadin.flow.internal.nodefeature.VirtualChildrenList;
 import com.vaadin.flow.server.VaadinService;
 import com.vaadin.flow.server.startup.CustomElementRegistry;
 
-import elemental.json.Json;
 import elemental.json.JsonArray;
 
 /**
@@ -59,10 +45,6 @@ import elemental.json.JsonArray;
  *
  */
 public class TemplateInitializer {
-    // {{propertyName}} or {{propertyName::event}}
-    private static final Pattern TWO_WAY_BINDING_PATTERN = Pattern
-            .compile("\\s*\\{\\{([^}:]*)(::[^}]*)?\\}\\}\\s*");
-
     private static final ConcurrentHashMap<TemplateParser, ReflectionCache<PolymerTemplate<?>, ParserData>> CACHE = new ConcurrentHashMap<>();
 
     private final PolymerTemplate<?> template;
@@ -71,102 +53,6 @@ public class TemplateInitializer {
     private final ParserData parserData;
 
     private final Map<String, Element> registeredElementIdToInjected = new HashMap<>();
-
-    private static class SubTemplateData {
-        private final String id;
-        private final String tag;
-        private final JsonArray path;
-
-        private SubTemplateData(String id, String tag, JsonArray path) {
-            this.id = id;
-            this.tag = tag;
-            this.path = path;
-        }
-    }
-
-    private interface InjectableFieldCunsumer {
-        void apply(Field field, String id, String tag);
-    }
-
-    /**
-     * Immutable parser data which may be stored in cache.
-     */
-    private static class ParserData {
-
-        private final Map<String, String> tagById;
-        private final Map<Field, String> idByField;
-
-        private final Set<String> twoWayBindingPaths;
-
-        private final Collection<SubTemplateData> subTemplates;
-
-        private ParserData(Map<Field, String> fields, Map<String, String> tags,
-                Set<String> twoWayBindings,
-                Collection<SubTemplateData> subTemplates) {
-            tagById = Collections.unmodifiableMap(tags);
-            idByField = Collections.unmodifiableMap(fields);
-            twoWayBindingPaths = Collections.unmodifiableSet(twoWayBindings);
-            this.subTemplates = Collections
-                    .unmodifiableCollection(subTemplates);
-        }
-
-        private void forEachInjectedField(InjectableFieldCunsumer consumer) {
-            idByField.forEach(
-                    (field, id) -> consumer.apply(field, id, tagById.get(id)));
-        }
-
-        private Set<String> getTwoWayBindingPaths() {
-            return twoWayBindingPaths;
-        }
-
-        private void forEachSubTemplate(
-                Consumer<SubTemplateData> dataConsumer) {
-            subTemplates.forEach(dataConsumer);
-        }
-    }
-
-    /**
-     * Mutable temporary parser data which is used to collected data only during
-     * initialization. The data will be discarded and not stored anywhere.
-     */
-    private static class ParserDataContext {
-        private final Map<String, String> tagById = new HashMap<>();
-        private final Map<Field, String> idByField = new HashMap<>();
-
-        private final Collection<SubTemplateData> subTemplates = new ArrayList<>();
-        private final Set<String> twoWayBindingPaths = new HashSet<>();
-        private final Set<String> notInjectableElementIds = new HashSet<>();
-
-        private org.jsoup.nodes.Element templateRoot;
-
-        private void addSubTemplate(String id, String tag, JsonArray path) {
-            subTemplates.add(new SubTemplateData(id, tag, path));
-        }
-
-        private void addNotInjectableId(String id) {
-            notInjectableElementIds.add(id);
-        }
-
-        private void addTwoWayBindingPath(String path) {
-            twoWayBindingPaths.add(path);
-        }
-
-        private void setTemplateRoot(org.jsoup.nodes.Element root) {
-            templateRoot = root;
-        }
-
-        private Optional<String> addTagName(String id, Field field) {
-            idByField.put(field, id);
-            Optional<String> tag = Optional
-                    .ofNullable(templateRoot.getElementById(id))
-                    .map(org.jsoup.nodes.Element::tagName);
-            if (tag.isPresent()) {
-                tagById.put(id, tag.get());
-            }
-            return tag;
-        }
-
-    }
 
     /**
      * Creates a new initializer instance.
@@ -192,37 +78,15 @@ public class TemplateInitializer {
             ReflectionCache<PolymerTemplate<?>, ParserData> cache = CACHE
                     .computeIfAbsent(parser,
                             analyzer -> new ReflectionCache<PolymerTemplate<?>, ParserData>(
-                                    clazz -> parseTemplate(clazz, analyzer)));
+                                    clazz -> new TemplateDataAnalyzer(clazz,
+                                            analyzer).parseTemplate()));
             data = cache.get(templateClass);
         }
         if (data == null) {
-            data = parseTemplate(templateClass, parser);
+            data = new TemplateDataAnalyzer(templateClass, parser)
+                    .parseTemplate();
         }
         parserData = data;
-    }
-
-    private ParserData readData(ParserDataContext context) {
-        return new ParserData(context.idByField, context.tagById,
-                context.twoWayBindingPaths, context.subTemplates);
-    }
-
-    private ParserData parseTemplate(Class<? extends PolymerTemplate<?>> clazz,
-            TemplateParser parser) {
-        org.jsoup.nodes.Element root = parser.getTemplateContent(templateClass,
-                getElement().getTag());
-        ParserDataContext data = new ParserDataContext();
-        data.setTemplateRoot(root);
-        Elements templates = root.getElementsByTag("template");
-        for (org.jsoup.nodes.Element element : templates) {
-            org.jsoup.nodes.Element parent = element.parent();
-            if (parent != null && getElement().getTag().equals(parent.id())) {
-                inspectCustomElements(element, element, data);
-
-                inspectTwoWayBindings(element, data);
-            }
-        }
-        collectInjectedIds(clazz, data);
-        return readData(data);
     }
 
     /**
@@ -243,78 +107,6 @@ public class TemplateInitializer {
         return parserData.getTwoWayBindingPaths();
     }
 
-    private void inspectCustomElements(org.jsoup.nodes.Element childElement,
-            org.jsoup.nodes.Element templateRoot, ParserDataContext data) {
-        if (isInsideTemplate(childElement, templateRoot)) {
-            storeNotInjectableElementId(childElement, data);
-        }
-
-        collectCustomElement(childElement, templateRoot, data);
-        childElement.children().forEach(
-                child -> inspectCustomElements(child, templateRoot, data));
-    }
-
-    private void storeNotInjectableElementId(org.jsoup.nodes.Element element,
-            ParserDataContext data) {
-        String id = element.id();
-        if (id != null && !id.isEmpty()) {
-            data.addNotInjectableId(id);
-        }
-    }
-
-    private void inspectTwoWayBindings(org.jsoup.nodes.Element element,
-            ParserDataContext data) {
-        Matcher matcher = TWO_WAY_BINDING_PATTERN.matcher("");
-        element.traverse(new NodeVisitor() {
-            @Override
-            public void head(Node node, int depth) {
-                // Two way bindings should only be in property bindings, not
-                // inside text content.
-                for (Attribute attribute : node.attributes()) {
-                    matcher.reset(attribute.getValue());
-                    if (matcher.matches()) {
-                        String path = matcher.group(1);
-                        data.addTwoWayBindingPath(path);
-                    }
-                }
-            }
-
-            @Override
-            public void tail(Node node, int depth) {
-                // Nop
-            }
-        });
-    }
-
-    private boolean isInsideTemplate(org.jsoup.nodes.Element element,
-            org.jsoup.nodes.Element templateRoot) {
-        if (element == templateRoot) {
-            return false;
-        }
-        if ("template".equalsIgnoreCase(element.tagName())) {
-            return true;
-        }
-        return isInsideTemplate(element.parent(), templateRoot);
-    }
-
-    private void collectCustomElement(org.jsoup.nodes.Element element,
-            org.jsoup.nodes.Element templateRoot, ParserDataContext data) {
-        String tag = element.tagName();
-
-        if (CustomElementRegistry.getInstance()
-                .isRegisteredCustomElement(tag)) {
-            if (isInsideTemplate(element, templateRoot)) {
-                throw new IllegalStateException("Couldn't parse the template: "
-                        + "sub-templates are not supported. Sub-template found: \n"
-                        + element);
-            }
-
-            String id = element.hasAttr("id") ? element.attr("id") : null;
-            JsonArray path = getPath(element, templateRoot);
-            data.addSubTemplate(id, tag, path);
-        }
-    }
-
     private void doRequestAttachCustomElement(String id, String tag,
             JsonArray path) {
         if (registeredElementIdToInjected.containsKey(id)) {
@@ -333,87 +125,9 @@ public class TemplateInitializer {
         CustomElementRegistry.getInstance().wrapElementIfNeeded(element);
     }
 
-    private JsonArray getPath(org.jsoup.nodes.Element element,
-            org.jsoup.nodes.Element templateRoot) {
-        List<Integer> path = new ArrayList<>();
-        org.jsoup.nodes.Element current = element;
-        while (!current.equals(templateRoot)) {
-            org.jsoup.nodes.Element parent = current.parent();
-            path.add(indexOf(parent, current));
-            current = parent;
-        }
-        JsonArray array = Json.createArray();
-        for (int i = 0; i < path.size(); i++) {
-            array.set(i, path.get(path.size() - i - 1));
-        }
-        return array;
-    }
-
-    /**
-     * Returns the index of the {@code child} in the collection of
-     * {@link org.jsoup.nodes.Element} children of the {@code parent} ignoring
-     * "style" elements.
-     * <p>
-     * "style" elements are handled differently depending on ES5/ES6. Also
-     * "style" tag can be moved on the top in the resulting client side DOM
-     * regardless of its initial position (e.g. Chrome does this).
-     *
-     * @param parent
-     *            the parent of the {@code child}
-     * @param child
-     *            the child element whose index is calculated
-     * @return the index of the {@code child} in the {@code parent}
-     */
-    private int indexOf(org.jsoup.nodes.Element parent,
-            org.jsoup.nodes.Element child) {
-        Elements children = parent.children();
-        int index = -1;
-        for (org.jsoup.nodes.Element nextChild : children) {
-            if (!"style".equals(nextChild.tagName())) {
-                index++;
-            }
-            if (nextChild.equals(child)) {
-                break;
-            }
-        }
-        return index;
-    }
-
     private ShadowRoot getShadowRoot() {
         return getElement().getShadowRoot()
                 .orElseGet(() -> getElement().attachShadow());
-    }
-
-    private void collectInjectedIds(Class<?> cls, ParserDataContext data) {
-        if (!AbstractTemplate.class.equals(cls.getSuperclass())) {
-            // Parent fields
-            collectInjectedIds(cls.getSuperclass(), data);
-        }
-
-        Stream.of(cls.getDeclaredFields()).filter(field -> !field.isSynthetic())
-                .forEach(field -> collectedInjectedId(field, data));
-    }
-
-    private void collectedInjectedId(Field field, ParserDataContext data) {
-        Optional<Id> idAnnotation = AnnotationReader.getAnnotationFor(field,
-                Id.class);
-        if (!idAnnotation.isPresent()) {
-            return;
-        }
-        String id = idAnnotation.get().value();
-        if (data.notInjectableElementIds.contains(id)) {
-            throw new IllegalStateException(String.format(
-                    "Class '%s' contains field '%s' annotated with @Id('%s'). "
-                            + "Corresponding element was found in a sub template, for which injection is not supported",
-                    templateClass.getName(), field.getName(), id));
-        }
-
-        if (!data.addTagName(id, field).isPresent()) {
-            throw new IllegalStateException(String.format(
-                    "There is no element with "
-                            + "id='%s' in the template file. Cannot map it using @%s",
-                    id, Id.class.getSimpleName()));
-        }
     }
 
     /* Map declared fields marked @Id */
@@ -540,7 +254,7 @@ public class TemplateInitializer {
 
     private void createSubTemplates() {
         parserData.forEachSubTemplate(data -> doRequestAttachCustomElement(
-                data.id, data.tag, data.path));
+                data.getId(), data.getTag(), data.getPath()));
     }
 
 }
